seperate controls and data
